Correct misinformation about this being an Arduino library
This is not an Arduino library. It is incompatible with the Arduino IDE or arduino-cli. This is a Particle library. Although Arduino libraries are compatible with Particle, Particle chose to make their system so that Particle libraries are usually not compatible with Arduino. Thus, calling this an "Arduino library" will only lead to confusion, frustration, and wasted time for Arduino users.